pyproject.toml
[project]
name = "usb-sentinel"
version = "0.1.0"
description = "LLM-integrated USB firewall system"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "USB Sentinel Contributors" }
]
keywords = ["usb", "security", "firewall", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Hardware :: Hardware Drivers",
]

dependencies = [
    "pyusb>=1.2.0",
    "pyyaml>=6.0",
    "anthropic>=0.20.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "sqlalchemy>=2.0.0",
    "websockets>=11.0",
    "pydantic>=2.0.0",
    "pyudev>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "httpx>=0.24.0",
]
local-llm = [
    "llama-cpp-python>=0.2.0",
]

[project.scripts]
usb-sentinel = "sentinel.cli:main"
sentinel-daemon = "sentinel.daemon:main"

[project.urls]
Homepage = "https://github.com/kase1111-hash/USB-Sentinel"
Repository = "https://github.com/kase1111-hash/USB-Sentinel"
Issues = "https://github.com/kase1111-hash/USB-Sentinel/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["sentinel"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["usb.*", "pyudev.*"]
ignore_missing_imports = true
